requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
attrs==23.1.0
    # via
    #   outcome
    #   trio
certifi==2023.7.22
    # via selenium
cffi==1.16.0
    # via trio
exceptiongroup==1.1.3
    # via
    #   trio
    #   trio-websocket
h11==0.14.0
    # via wsproto
idna==3.4
    # via trio
outcome==1.2.0
    # via trio
pillow==10.1.0
    # via ttkthemes
pycparser==2.21
    # via cffi
pysocks==1.7.1
    # via urllib3
pyyaml==6.0.1
    # via -r requirements.in
selenium==4.14.0
    # via -r requirements.in
sniffio==1.3.0
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.22.2
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
ttkthemes==3.2.2
    # via -r requirements.in
urllib3[socks]==2.0.6
    # via selenium
wsproto==1.2.0
    # via trio-websocket
